Added react icons, Splitted restaurant page into multiple files, Added Instagram, Info Icons.
