+++ package.json
@@ -34,7 +34,8 @@
         "framer-motion": "^4.1.17",
         "next": "latest",
         "react": "^17.0.2",
-        "react-dom": "^17.0.2"
+        "react-dom": "^17.0.2",
+        "react-icons": "^4.2.0"
     },
     "devDependencies": {
         "@testing-library/react": "^11.2.5",
